Refactor TrinoBridge and Utils.
